requirements.txt
dj-database-url==0.4.1
Django==1.9.5
django-dotenv==1.4.1
django-localflavor==1.2
psycopg2==2.6.1
wheel==0.24.0
